Refactor y mejora de clases: se ajustan tipos de datos, se añaden validaciones y se corrigen mensajes de error

+++ src/main/kotlin/ar/edu/unsam/phm/dto/PasajeroDTO.kt
@@ -22,7 +22,7 @@ class InformacionPasajeroDTO(
     val nombre:String,
     val apellido:String,
     val telefono:String,
-    val saldo:Long,
+    val saldo: Double,
     val listaAmigos:List<TarjetaAmigoDTO>
 )
